// src/services/clients/clients.d.ts
export interface IAdminUsersClient extends Service {
  getUsers(
    query: string,
    page?: number,
    pageSize?: number
  ): Promise<IJsonResponse>;

  getUser(id: number): Promise<IJsonResponse>;

  deleteUser(id: number): Promise<IJsonResponse>;

  updateUser(
    id: number,
    firstName: string,
    lastName: string,
    email: string,
    phoneNumber: string,
    roleType: string,
    permissions: number[]
  ): Promise<IJsonResponse>;

  disableUser(id: number): Promise<IJsonResponse>;

  enableUser(id: number): Promise<IJsonResponse>;

  createAdminUser(
    firstName: string,
    lastName: string,
    email: string,
    phoneNumber: string,
    password: string,
    roleType: string,
    permissions: number[]
  ): Promise<IJsonResponse>;

  getCurrentAdmin(): Promise<IJsonResponse>;

}